requirements.txt
psycopg2-binary
customtkinter
